refactor(es): Relocate the shared RoundTripper stack into esclient
Move GetHTTPRoundTripper, getBodyFixRoundTripper, customHeadersRoundTripper, and
the auth_helper init functions from clientbuilder into the esclient package
(RFC 0006 §6.4/§6.1). clientbuilder's data-plane callers now call
esclient.GetHTTPRoundTripper; the relocated unit tests move into
esclient/roundtripper_test.go. Pure code move, zero behavior change: the stack
uses only config/internal/auth/extensionauth/tls/http (no olivere), so esclient
stays olivere-free and can host the transport that both planes share. This lets
the admin esclient.Client build its base through the full stack next.
