+++ internal/storage/elasticsearch/esclient/auth_helper.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package clientbuilder
+package esclient
 
 import (
 	"context"
